work on LoadModel() SaveModel(); more flexible correlation computation (float as 'overlap' input, allows normalization etc.)

+++ src/MyMediaLite/Correlation/BidirectionalConditionalProbability.cs
@@ -66,7 +66,7 @@ namespace MyMediaLite.Correlation
 		}
 
 		///
-		protected override float ComputeCorrelationFromOverlap(uint overlap, int count_x, int count_y)
+		protected override float ComputeCorrelationFromOverlap(float overlap, int count_x, int count_y)
 		{
 			if (count_x == 0 || count_y == 0)
 				return 0.0f;
