+++ src/MyMediaLite/Correlation/BinaryCosine.cs
@@ -53,7 +53,7 @@ namespace MyMediaLite.Correlation
 		}
 
 		///
-		protected override float ComputeCorrelationFromOverlap(uint overlap, int count_x, int count_y)
+		protected override float ComputeCorrelationFromOverlap(float overlap, int count_x, int count_y)
 		{
 			long size_product = count_x * count_y;
 			if (size_product > 0.0f)
